+++ router/core/errors.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
 	"net"
 	"net/http"
 
 	"github.com/hashicorp/go-multierror"
 	rErrors "github.com/wundergraph/cosmo/router/internal/errors"
-	"github.com/wundergraph/cosmo/router/internal/persistedoperation"
 	"github.com/wundergraph/cosmo/router/internal/unique"
 	"github.com/wundergraph/cosmo/router/pkg/pubsub"
 	rtrace "github.com/wundergraph/cosmo/router/pkg/trace"
@@ -115,7 +115,7 @@ func trackFinalResponseError(ctx context.Context, err error) {
 
 func getAggregatedSubgraphErrorCodes(err error) []string {
 
-	if unwrapped, ok := err.(MultiError); ok {
+	if unwrapped, ok := err.(multiError); ok {
 
 		errs := unwrapped.Unwrap()
 
@@ -144,7 +144,7 @@ func getSubgraphNames(ds []resolve.DataSourceInfo) []string {
 
 func getAggregatedSubgraphServiceNames(err error) []string {
 
-	if unwrapped, ok := err.(MultiError); ok {
+	if unwrapped, ok := err.(multiError); ok {
 
 		errs := unwrapped.Unwrap()
 
@@ -281,7 +281,8 @@ func writeOperationError(r *http.Request, w http.ResponseWriter, requestLogger *
 	case errors.As(err, &httpErr):
 		writeRequestErrors(r, w, httpErr.StatusCode(), requestErrorsFromHttpError(httpErr), requestLogger)
 	case errors.As(err, &poNotFoundErr):
-		writeRequestErrors(r, w, http.StatusBadRequest, graphqlerrors.RequestErrorsFromError(errors.New("persisted Query not found")), requestLogger)
+		newErr := NewHttpGraphqlError("persisted query not found", "PERSISTED_QUERY_NOT_FOUND", http.StatusOK)
+		writeRequestErrors(r, w, http.StatusOK, requestErrorsFromHttpError(newErr), requestLogger)
 	case errors.As(err, &reportErr):
 		report := reportErr.Report()
 		logInternalErrorsFromReport(reportErr.Report(), requestLogger)