+++ router/CHANGELOG.md
@@ -4,6 +4,30 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.140.2](https://github.com/wundergraph/cosmo/compare/router@0.140.1...router@0.140.2) (2024-11-08)
+
+### Bug Fixes
+
+* json parsing and input templates rendering ([#1350](https://github.com/wundergraph/cosmo/issues/1350)) ([578a408](https://github.com/wundergraph/cosmo/commit/578a4085862b3149dbbfb10362aff118c62dfff2)) (@jensneuse)
+* **router:** don't flush all redis keys while shutdown ([#1349](https://github.com/wundergraph/cosmo/issues/1349)) ([9f37105](https://github.com/wundergraph/cosmo/commit/9f371059ec9608fce5f9da8c72246604b46fd34d)) (@git-hulk)
+
+## [0.140.1](https://github.com/wundergraph/cosmo/compare/router@0.140.0...router@0.140.1) (2024-11-08)
+
+### Bug Fixes
+
+* default metric temporality selector and override the selector for the cloud endpoint ([#1331](https://github.com/wundergraph/cosmo/issues/1331)) ([2a292ea](https://github.com/wundergraph/cosmo/commit/2a292eaec75624282becec7b84c9b57ab37a5f50)) (@JivusAyrus)
+* **router:** l2 cache uses client name in key, persisted operation not found returns 200 ([#1351](https://github.com/wundergraph/cosmo/issues/1351)) ([34cf5b5](https://github.com/wundergraph/cosmo/commit/34cf5b5c7f58b698c032c159c24cdda1cfe1bbdf)) (@df-wg)
+
+# [0.140.0](https://github.com/wundergraph/cosmo/compare/router@0.139.2...router@0.140.0) (2024-11-07)
+
+### Features
+
+* add apq back in ([#1346](https://github.com/wundergraph/cosmo/issues/1346)) ([61c0d9b](https://github.com/wundergraph/cosmo/commit/61c0d9b9ac383ad744d47a461fc7ac9b07c9c981)) (@df-wg)
+
+## [0.139.2](https://github.com/wundergraph/cosmo/compare/router@0.139.1...router@0.139.2) (2024-11-07)
+
+**Note:** Version bump only for package router
+
 ## [0.139.1](https://github.com/wundergraph/cosmo/compare/router@0.139.0...router@0.139.1) (2024-11-06)
 
 ### Bug Fixes